Update main-v4.js
You correctly implement the proof of work in the mining function, however, it is not enforced in the isChainValid() function. This means that block tampering attack can still easily be mounted by re-computing hashes of the victim block and all blocks afterwards. 
In such an attack, the difficulty can be ignored and the chain will still be valid since you don't check whether each hash actually has the right preceding zeroes.